+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New Features
 
  - PR #529 Add debug logging and fix multithreaded replay benchmark
+ - PR #560 Remove deprecated `get/set_default_resource` APIs
 
 ## Improvements